Add support for conformances in type declarations
