+++ src/Swift.Bindings/src/Model/TypeDecl/TypeDecl.cs
@@ -27,5 +27,10 @@ namespace BindingsGeneration
         /// Types declarations within the base declaration.
         /// </summary>
         public required List<TypeDecl> Types { get; set; }
+
+        /// <summary>
+        /// Protocol conformances.
+        /// </summary>
+        public required List<Conformance> Conformances { get; set; }
     }
 }
